requirements.txt
appdirs==1.4.4
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
identify==1.5.11
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
libmagic==1.0
lxml==4.6.2
mccabe==0.6.1
nodeenv==1.5.0
numpy==1.19.4
opencv-python==4.5.1.48
Pillow==8.1.0
pre-commit==2.9.3
pycodestyle==2.6.0
PyPDF2==1.26.0
pytesseract==0.3.7
python-docx==0.8.10
python-magic==0.4.18
PyYAML==5.3.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
